Commenting and Tweaking
Still quite a bit of bugs, noteably children getting one extra element in there diet when conceived, leading to very large Strings of nutrients the further into a generation it goes. Also there was this one time there was a shitton of creatures with 0 max hunger and current hunger with a diet of the entire alphabet... wacky things.